requirements.txt
einops >= 0.3.0
hydra-core >= 1.1.0
monai  >= 0.7.0
nibabel >= 3.2.2
opencv-contrib-python >= 4.5.4.60
opencv-python >= 4.5.5.62
pandas >= 1.3.5
Pillow >= 9.0.1
pip >= 22.3.1
pydicom  >= 2.2.2
pyradiomics  >= 3.0.1
pytorch-ignite >= 0.4.9
pytorch-model-summary >= 0.1.2
pytorch-pretrained-vit >= 0.0.7
pytorch-warmup >= 0.0.4
PyYAML   >= 5.4.1
SimpleITK  >= 2.1.1
tensorboard  >= 2.1.1
tensorboardX >= 2.5.1
timm >= 0.4.12
torch >= 1.9.0+cu111
torchsummary >= 1.5.1
torchvision >= 0.10.0+cu111
tqdm >= 4.64.1
vit-pytorch >= 0.21.1
vtk >= 9.1.0

